pyproject.toml
[tool.poetry]
name = "zigral"
version = "3.0.0"
description = "Zigral - AI-powered sales prospecting and outreach automation platform"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
openai = "^1.3.0"
playwright = "^1.39.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tortoise-orm = "^0.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing" 